+++ nkululeko/explore.py
@@ -42,7 +42,7 @@ def main():
     )
     parser.add_argument("--config", default="exp.ini", help="The base configuration")
     args = parser.parse_args()
-    config_file = args.config if args.config is not None else "exp.ini"
+    config_file = args.config
 
     try:
         if not Path(config_file).is_file():