+++ nkululeko/multidb.py
@@ -26,16 +26,13 @@ from nkululeko.nkululeko import doit as nkulu
 from nkululeko.utils.errors import NkululukoError
 
 
-def main(src_dir):
+def main():
     parser = argparse.ArgumentParser(
         description="Call the nkululeko MULTIDB framework."
     )
     parser.add_argument("--config", default="exp.ini", help="The base configuration")
     args = parser.parse_args()
-    if args.config is not None:
-        config_file = args.config
-    else:
-        config_file = f"{src_dir}/exp.ini"
+    config_file = args.config
 
     # test if the configuration file exists
     if not os.path.isfile(config_file):
@@ -201,5 +198,4 @@ def plot_heatmap(results, last_epochs, labels, name, config, datasets):
 
 
 if __name__ == "__main__":
-    cwd = os.path.dirname(os.path.abspath(__file__))
-    main(cwd)  # sys.argv[1])
+    main()